Remove temp file created
Explicitly removes the temporary file created for randomising the note name.

On my system (Elementary 5.1) Based on Ubuntu 18.04 LTS. I am using the .md file extension added into the todo config file as suggested.  When creating a note, it creates two files, one with the.md extension and one with no extension so I have to rm the random generated file. Not sure if this is replicated on other systems, but explicitly removing it should not cause a problem if it doesn't exist.